Temporarily remove arm64 build
This is done to avoid being blocked by a regression introduced by
27b5ebf61818749b3568354c64a8ec2d9cd5ecca where compilation fails
with:
./drivers/firmware/efi/libstub/lib.a(arm64-stub.stub.o): In function `handle_kernel_image':
./drivers/firmware/efi/libstub/arm64-stub.c:63: undefined reference to `__efistub__GLOBAL_OFFSET_TABLE_'
/usr/bin/aarch64-linux-gnu-ld: ./drivers/firmware/efi/libstub/lib.a(arm64-stub.stub.o): relocation R_AARCH64_ADR_PREL_PG_HI21 against external symbol `__efistub__GLOBAL_OFFSET_TABLE_' can not be used when making a shared object; recompile with -fPIC

This is mysteriously addressed as fixed on the Debian bugtracker (see
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=918188), witha commit
name that is not present in the log and might have been cherry picked
by them.

It will need further investigation, but for now, let's make the release
work.